Define the comparator as a separate class.
